refactor: rename retrofit package to okhttp
Move ContinuationCallback, OkHttpBlockerNetwork, and its test from
the retrofit subpackage to okhttp, completing the Retrofit removal.
