+++ core/network/src/main/kotlin/com/merxury/blocker/core/network/okhttp/OkHttpBlockerNetwork.kt
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.merxury.blocker.core.network.retrofit
+package com.merxury.blocker.core.network.okhttp
 
 import com.merxury.blocker.core.model.preference.RuleServerProvider
 import com.merxury.blocker.core.model.preference.RuleServerProvider.GITHUB
